Update updateEncoderCount()
Brought linearConv functionality into updateEncoderCount(), modified this functionality, light refactoring

+++ prototype_ring_design.cydsn/i2cFunctions.h
@@ -16,7 +16,7 @@
     typedef struct
     {
         MotorAddress addr;
-    	int counter;
+    	uint8 counter;
     	int index;
         int calibrationLength;
         int lineLength;
